Refactor app to use feature folders.
