--- .babelrc
@@ -1,3 +1,0 @@
-{
-  "plugins": ["transform-class-properties"]
-}
