+++ src/js/app/components/DeleteModal.js
