+++ src/js/app/components/Map.js
