+++ src/js/auth/reducers/authReducer.js
@@ -1,0 +1,6 @@
+import { handleActions } from 'redux-actions';
+import UserData from '../records/UserData';
+
+const initialState = new UserData();
+
+export default handleActions({}, initialState);
