+++ src/js/bike-racks/constants/bikeRacksReduxConfig.js
@@ -1,0 +1,3 @@
+export const name = 'bikeracks';
+export const path = 'admin/bike-racks/';
+export const indexFn = bikerack => bikerack.get('id');
