+++ src/js/bike-racks/records/BikeRack.js
@@ -1,0 +1,14 @@
+import { Record } from 'immutable';
+
+export default Record(
+  {
+    id: null,
+    bikeCount: null,
+    name: null,
+    description: null,
+    checkInArea: null,
+    lat: null,
+    lon: null,
+  },
+  'BikeRack'
+);
