+++ src/js/bikes/actions/bikeActions.js
@@ -1,0 +1,7 @@
+import Bike from '../records/Bike';
+import { name, path, indexFn } from '../constants/bikeReduxConfig';
+import { all } from 'common/all';
+
+const { actions } = all({ name, path, record: Bike, indexFn });
+
+export default actions;
