+++ src/js/bikes/constants/BikeActionTypes.js
@@ -1,0 +1,2 @@
+export const BIKE_EDITOR_OPEN = 'BIKE_EDITOR_OPEN';
+export const BIKE_EDITOR_CLOSE = 'BIKE_EDITOR_CLOSE';
