+++ src/js/bikes/constants/bikeReduxConfig.js
@@ -1,0 +1,3 @@
+export const name = 'bikes';
+export const path = 'admin/bikes/';
+export const indexFn = bike => bike.get('id');
