+++ src/js/bikes/records/Bike.js
@@ -1,0 +1,14 @@
+import { Record } from 'immutable';
+
+export default Record(
+  {
+    id: null,
+    lock: null,
+    visible: true,
+    currentRental: null,
+    previousRental: null,
+    lat: null,
+    lon: null,
+  },
+  'Bike'
+);
