--- src/js/components/BikeRacks/index.js
@@ -1,3 +1,0 @@
-import BikeRacksView from './BikeRacksView';
-
-export default BikeRacksView;
