--- src/js/components/Bikes/index.js
@@ -1,3 +1,0 @@
-import BikesView from './BikesView';
-
-export default BikesView;
