--- src/js/components/Home/index.js
@@ -1,3 +1,0 @@
-import HomeView from './HomeView';
-
-export default HomeView;
