--- src/js/components/Nav/index.js
@@ -1,3 +1,0 @@
-import NavView from './NavView';
-
-export default NavView;
