--- src/js/redux/auth/reducer.js
@@ -1,6 +1,0 @@
-import { handleActions } from 'redux-actions';
-import UserData from '../../data/records/UserData';
-
-const initialState = new UserData();
-
-export default handleActions({}, initialState);
